# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Deployer API
  version: v1
  description: >
    Machine-readable reference for the Deployer admin API and runner agent API.
    Admin routes are intended to be protected by an upstream authorization
    gateway; Deployer does not implement local admin password login. Browser
    write requests also send X-Deployer-CSRF: "1". Agent routes use bearer
    tokens created through the runner API. Private control-plane routes use
    scoped Deployer service tokens and are not customer-facing endpoints.
servers:
  - url: https://deployer.example.com
    description: Example production URL
  - url: http://127.0.0.1:9090
    description: Default local URL
...
paths:
...
  /api/projects:
    get:
      tags: [projects]
      summary: List projects
      operationId: listProjects
      responses:
        "200":
          description: Projects with their latest build when available.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags: [projects]
      summary: Create a project
      operationId: createProject
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
            examples:
              filesProject:
                summary: Files-mode project
                value:
                  name: Dashboard
                  repo_path: /srv/repos/dashboard
                  deploy_dir: /srv/apps/dashboard
                  deploy_mode: files
                  runner_id: 1
                  preserve: uploads,.env
                  permissions: '{"owner":"www-data:www-data","dirs":{"storage":"0750"}}'
              dockerProject:
                summary: Docker-mode project
                value:
                  name: API
                  repo_path: /srv/repos/api
                  dockerfile_path: Dockerfile
                  compose_file: docker-compose.yml
                  compose_services: app worker
                  image_name: registry.example.com/api
                  deploy_dir: /srv/apps/api
                  deploy_mode: docker
                  runner_id: 1
                  health_url: https://api.example.com/health
      responses:
        "201":
          description: Created project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/Error"
  /api/projects/import:
    post:
      tags: [projects]
      summary: Import a project from a public-safe export envelope
      operationId: importProject
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectExport"
      responses:
        "201":
          description: Imported project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/Error"
  /api/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [projects]
      summary: Get a project
      operationId: getProject
      responses:
        "200":
          description: Project details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [projects]
      summary: Update a project
      operationId: updateProject
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/Error"
    delete:
      tags: [projects]
      summary: Delete a project
      operationId: deleteProject
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      responses:
        "204":
          description: Project deleted.
...
  /api/builds/{buildId}:
    parameters:
      - $ref: "#/components/parameters/BuildId"
    get:
      tags: [builds]
      summary: Get a build
      operationId: getBuild
      responses:
        "200":
          description: Build details and persisted log.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/runners:
    get:
      tags: [runners]
      summary: List runners
      operationId: listRunners
      responses:
        "200":
          description: Runners without token values.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Runner"
    post:
      tags: [runners]
      summary: Create a runner
      operationId: createRunner
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerInput"
            examples:
              createRunner:
                value:
                  name: production-vps
                  labels: linux,docker
      responses:
        "201":
          description: Created runner. The token is returned only once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
              examples:
                runnerCreated:
                  value:
                    id: 1
                    name: production-vps
                    token: dep_1234567890_example
                    labels: linux,docker
                    status: offline
                    created_at: "2026-06-16T12:00:00Z"
        "400":
          $ref: "#/components/responses/Error"
...
  /api/service-tokens:
    get:
      tags: [service-tokens]
      summary: List service tokens without their secret values
      operationId: listServiceTokens
      responses:
        "200":
          description: Service token metadata.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceToken"
    post:
      tags: [service-tokens]
      summary: Create a scoped service token
      operationId: createServiceToken
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTokenInput"
      responses:
        "201":
          description: Token metadata with the secret returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceToken"
...
  /api/hosting-runners:
    get:
      tags: [runners]
      summary: List dedicated hosting runners
      operationId: listHostingRunners
      responses:
        "200":
          description: Hosting runner metadata without credentials.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/HostingRunnerAdmin"}
    post:
      tags: [runners]
      summary: Register a dedicated hosting runner
      operationId: createHostingRunner
      parameters:
        - $ref: "#/components/parameters/CSRFHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/HostingRunnerInput"}
      responses:
        "201":
          description: Runner with one-time dedicated hosting credential.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HostingRunnerAdmin"}
        "400": {$ref: "#/components/responses/Error"}
...
  /api/internal/v1/capabilities:
    get:
      tags: [internal]
      summary: Negotiate hosting manifest, runner, runtime, resource, and failure-code capabilities
      operationId: getInternalCapabilities
      security: [{serviceBearerAuth: []}]
      responses:
        "200":
          description: Hosting execution capabilities.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HostingCapabilities"}
  /api/internal/v1/runners:
    get:
      tags: [internal]
      summary: List hosting-only runner capabilities and free capacity
      operationId: listInternalHostingRunners
      security: [{serviceBearerAuth: []}]
      responses:
        "200":
          description: Hosting runners; legacy trusted-admin runners are excluded.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/HostingRunnerState"}
...
  /api/hosting-agent/v1/heartbeat:
    post:
      tags: [hosting-agent]
      summary: Publish hosting runner liveness, free capacity, drain state, and negotiated versions
      operationId: heartbeatHostingAgent
      security: [{hostingAgentBearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/HostingHeartbeat"}
      responses:
        "200":
          description: Heartbeat accepted with the authoritative release-retention set for this runner.
          content:
            application/json:
              schema:
                type: object
                required: [retained_releases]
                properties:
                  retained_releases:
                    type: array
                    items: {$ref: "#/components/schemas/HostingRetainedRelease"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
  /api/hosting-agent/v1/poll:
    post:
      tags: [hosting-agent]
      summary: Atomically claim one job assigned to this hosting runner
      operationId: pollHostingAgent
      security: [{hostingAgentBearerAuth: []}]
      responses:
        "200":
          description: Lease-bound generated workload recipe and opaque secret references.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HostingClaimedJob"}
        "204": {description: No assigned job is available.}
...
  /api/agent/poll:
    get:
      tags: [agent]
      summary: Claim the next pending job for a runner
      operationId: pollAgentJob
      security:
        - agentBearerAuth: []
      responses:
        "200":
          description: A job was claimed for this runner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "204":
          description: No pending job was available before the long-poll timeout.
        "401":
          $ref: "#/components/responses/Error"
...
  /api/agent/heartbeat:
    post:
      tags: [agent]
      summary: Update runner heartbeat
      operationId: heartbeatAgent
      security:
        - agentBearerAuth: []
      responses:
        "200":
          description: Heartbeat accepted.
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    agentBearerAuth:
      type: http
      scheme: bearer
    serviceBearerAuth:
      type: http
      scheme: bearer
      description: Hashed scoped service token. Supported scopes are deployments:read, deployments:write, projects:write, and hosting:admin.
    hostingAgentBearerAuth:
      type: http
      scheme: bearer
      description: Dedicated hashed hosting-runner credential; never interchangeable with a legacy runner or service token.
  parameters:
    CSRFHeader:
      name: X-Deployer-CSRF
      in: header
      required: true
      schema:
        type: string
        enum: ["1"]
      description: Required for state-changing browser/admin API requests.
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    BuildId:
      name: buildId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Error:
      description: JSON error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    ProjectInput:
      type: object
      required: [name, repo_path, deploy_dir]
      properties:
        name:
          type: string
        repo_path:
          type: string
        dockerfile_path:
          type: string
          default: Dockerfile
        compose_file:
          type: string
          default: docker-compose.yml
        image_name:
          type: string
        ssh_host:
          type: string
        deploy_dir:
          type: string
        health_url:
          type: string
        health_container:
          type: string
        build_args:
          type: object
          additionalProperties:
            type: string
        compose_services:
          type: string
        git_pull_before_build:
          type: boolean
        runner_id:
          type: integer
          format: int64
        deploy_mode:
          type: string
          enum: [docker, files]
          default: docker
        post_deploy:
          type: string
        permissions:
          type: string
          description: JSON string containing permissions configuration.
        preserve:
          type: string
...
    ProjectExport:
      type: object
      required: [kind, version, exported_at, project]
      properties:
        kind:
          type: string
          const: deployer.project
        version:
          type: integer
          const: 1
        exported_at:
          type: string
          format: date-time
        project:
          $ref: "#/components/schemas/ProjectExportConfig"
        omitted_fields:
          type: array
          items:
            type: string
          description: Fields intentionally omitted from the export, such as runner_id, post_deploy, or sensitive build args.
...
    Project:
      allOf:
        - $ref: "#/components/schemas/ProjectInput"
        - type: object
          required: [id, created_at]
          properties:
            id:
              type: integer
              format: int64
            created_at:
              type: string
              format: date-time
            last_build:
              $ref: "#/components/schemas/Build"
...